microservices/storage-account-service/requirements.txt
uvicorn
fastapi
pydantic
azure-storage-file-share